pyproject.toml
[tool.poetry]
name = "google-keep-to-docs"
version = "0.1.0"
description = "A script to fetch Google Keep notes and write them to Google Docs."
authors = ["Your Name <your_email@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yourusername/google-keep-to-docs"

[tool.poetry.dependencies]
python = "^3.8"
gkeepapi = "^0.14.0" # Unofficial library for Google Keep
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
google-api-python-client = "^2.94.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
